requirements.txt
pytesseract
nltk
pandas
joblib
pdfplumber
opencv-python-headless
Pillow
PyMuPDF
xgboost